fix: dedupe import and SVG helpers to clear Sonar duplication gate
- Add importArticleShared.ts for distrib/outrank payload path validation and article helpers
- Add imageSvgShared.ts for shared logo/text SVG utilities used by social and blog generators
- Add workflow-level contents: read to update-org-readme.yml
